:white_check_mark: :zap: various tests and perf
Buffer does not need to be concerned about garbage data. Vec ensures
that memory is cleared beforehand, which is a waste of cycles. The new
Buffer doesn't clean up memory.

Buffer moved to its own crate as it uses unsafe extensively.

Add tests for NBT.

Add tests for insecure crypto.